home-server/services/transcription/requirements.txt
faster-whisper>=1.0.0
paho-mqtt>=2.0.0
numpy>=1.24.0
